sistemaDeNotas/requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3
Django==2.1
django-adminlte2==0.4.1
django-bootstrap3==10.0.1
django-guardian==1.4.9
django-mysql==2.4.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
idna==2.7
jdcal==1.4
Jinja2==2.10
jsonfield==2.0.2
MarkupSafe==1.0
mo-future==2.18.18240
moz-sql-parser==2.18.18240
mysql-connector-python==8.0.12
mysqlclient==1.3.13
numpy==1.13.3
olefile==0.45.1
openpyxl==2.5.5
pandas==0.23.4
Pillow==5.2.0
protobuf==3.6.0
pycparser==2.18
PyMySQL==0.9.2
pyparsing==2.2.0
Pypubsub==4.0.0
python-dateutil==2.7.3
python-memcached==1.59
pytz==2018.5
reportlab==3.5.4
requests==2.19.1
six==1.11.0
sqlparse==0.2.4
urllib3==1.23
XlsxWriter==1.0.5
xlwt==1.3.0
django-polymorphic
django-ckeditor